+++ partiql-ast/src/main/kotlin/org/partiql/ast/helpers/ToLegacyAst.kt
@@ -39,7 +39,6 @@ import org.partiql.lang.ast.IsListParenthesizedMeta
 import org.partiql.lang.ast.IsValuesExprMeta
 import org.partiql.lang.ast.Meta
 import org.partiql.lang.domains.PartiqlAst
-import org.partiql.pig.runtime.SymbolPrimitive
 import org.partiql.value.DateValue
 import org.partiql.value.MissingValue
 import org.partiql.value.PartiQLValueExperimental
@@ -178,7 +177,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
     }
 
     override fun visitTableDefinitionColumn(node: TableDefinition.Column, ctx: Ctx) = translate(node) { metas ->
-        val name = stringToDefnid(node.name)
+        val name = defnid(node.name)
         val type = visitType(node.type, ctx)
         val constraints = node.constraints.translate<PartiqlAst.ColumnConstraint>(ctx)
         columnDeclaration(name, type, constraints, metas)
@@ -188,7 +187,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         node: TableDefinition.Column.Constraint,
         ctx: Ctx,
     ) = translate(node) { metas ->
-        val name = node.name?.let { stringToDefnid(it) }
+        val name = node.name?.let { defnid(it) }
         val def = when (node.body) {
             is TableDefinition.Column.Constraint.Body.Check -> {
                 throw IllegalArgumentException("PIG AST does not support CHECK (<expr>) constraint")
@@ -293,16 +292,16 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         if (node.function is Identifier.Qualified) {
             error("Qualified identifiers are not allowed in legacy AST `call` function identifiers")
         }
-        val funcName = stringToDefnid((node.function as Identifier.Symbol).symbol)
+        val funcName = (node.function as Identifier.Symbol).symbol.lowercase()
         val args = node.args.translate<PartiqlAst.Expr>(ctx)
-        call(funcName.lowercase(), args, metas)
+        call(defnid(funcName), args, metas)
     }
 
     override fun visitExprAgg(node: Expr.Agg, ctx: Ctx) = translate(node) { metas ->
         val setq = node.setq?.toLegacySetQuantifier() ?: all()
         // Legacy AST translates COUNT(*) to COUNT(1)
         if (node.function is Identifier.Symbol && (node.function as Identifier.Symbol).symbol == "COUNT_STAR") {
-            val funId = stringToDefnid("count")
+            val funId = defnid("count")
             return callAgg(setq, funId, lit(ionInt(1)), metas)
         }
         // Default Case
@@ -313,9 +312,9 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
             error("Qualified identifiers are not allowed in legacy AST `call_agg` function identifiers")
         }
         // Legacy parser/ast always inserts ALL quantifier
-        val funcId = stringToDefnid((node.function as Identifier.Symbol).symbol)
+        val funcName = (node.function as Identifier.Symbol).symbol.lowercase()
         val arg = visitExpr(node.args[0], ctx)
-        callAgg(setq, funcId.lowercase(), arg, metas)
+        callAgg(setq, defnid(funcName), arg, metas)
     }
 
     override fun visitExprUnary(node: Expr.Unary, ctx: Ctx) = translate(node) { metas ->
@@ -527,7 +526,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         val start = visitOrNull<PartiqlAst.Expr>(node.start, ctx)
         val length = visitOrNull<PartiqlAst.Expr>(node.length, ctx)
         val operands = listOfNotNull(value, start, length)
-        val funId = stringToDefnid("substring")
+        val funId = defnid("substring")
         call(funId, operands, metas)
     }
 
@@ -535,7 +534,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         val lhs = visitExpr(node.lhs, ctx)
         val rhs = visitExpr(node.rhs, ctx)
         val operands = listOf(lhs, rhs)
-        val funId = stringToDefnid("position")
+        val funId = defnid("position")
         call(funId, operands, metas)
     }
 
@@ -548,7 +547,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         if (spec != null) operands.add(lit(ionSymbol(spec)))
         if (chars != null) operands.add(chars)
         operands.add(value)
-        val funId = stringToDefnid("trim")
+        val funId = defnid("trim")
         call(funId, operands, metas)
     }
 
@@ -558,7 +557,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         val start = visitExpr(node.start, ctx)
         val length = visitOrNull<PartiqlAst.Expr>(node.length, ctx)
         val operands = listOfNotNull(value, overlay, start, length)
-        val funId = stringToDefnid("overlay")
+        val funId = defnid("overlay")
         call(funId, operands, metas)
     }
 
@@ -566,7 +565,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         val field = node.field.toLegacyDatetimePart()
         val source = visitExpr(node.source, ctx)
         val operands = listOf(field, source)
-        val funId = stringToDefnid("extract")
+        val funId = defnid("extract")
         call(funId, operands, metas)
     }
 
@@ -593,7 +592,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         val lhs = visitExpr(node.lhs, ctx)
         val rhs = visitExpr(node.rhs, ctx)
         val operands = listOf(field, lhs, rhs)
-        val funId = stringToDefnid("date_add")
+        val funId = defnid("date_add")
         call(funId, operands, metas)
     }
 
@@ -602,7 +601,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         val lhs = visitExpr(node.lhs, ctx)
         val rhs = visitExpr(node.rhs, ctx)
         val operands = listOf(field, lhs, rhs)
-        val funId = stringToDefnid("date_diff")
+        val funId = defnid("date_diff")
         call(funId, operands, metas)
     }
 
@@ -633,10 +632,10 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
     }
 
     override fun visitExprWindow(node: Expr.Window, ctx: Ctx) = translate(node) { metas ->
+        val funcName = node.function.name.lowercase()
         val over = visitExprWindowOver(node.over, ctx)
         val args = listOfNotNull(node.expression, node.offset, node.default).translate<PartiqlAst.Expr>(ctx)
-        val funId = stringToDefnid(node.function.name)
-        callWindow(funId.lowercase(), over, args, metas)
+        callWindow(defnid(funcName), over, args, metas)
     }
 
     override fun visitExprWindowOver(node: Expr.Window.Over, ctx: Ctx) = translate(node) { metas ->
@@ -709,7 +708,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
     override fun visitSelectProjectItemExpression(node: Select.Project.Item.Expression, ctx: Ctx) =
         translate(node) { metas ->
             val expr = visitExpr(node.expr, ctx)
-            val alias = node.asAlias?.let { stringToDefnid(it.symbol) }
+            val alias = node.asAlias?.let { defnid(it.symbol) }
             projectExpr(expr, alias, metas)
         }
 
@@ -727,9 +726,9 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
     override fun visitFrom(node: From, ctx: Ctx) = super.visitFrom(node, ctx) as PartiqlAst.FromSource
     override fun visitFromValue(node: From.Value, ctx: Ctx) = translate(node) { metas ->
         val expr = visitExpr(node.expr, ctx)
-        val asAlias = node.asAlias?.let { stringToDefnid(it.symbol) }
-        val atAlias = node.atAlias?.let { stringToDefnid(it.symbol) }
-        val byAlias = node.byAlias?.let { stringToDefnid(it.symbol) }
+        val asAlias = node.asAlias?.let { defnid(it.symbol) }
+        val atAlias = node.atAlias?.let { defnid(it.symbol) }
+        val byAlias = node.byAlias?.let { defnid(it.symbol) }
         when (node.type) {
             From.Value.Type.SCAN -> scan(expr, asAlias, atAlias, byAlias, metas)
             From.Value.Type.UNPIVOT -> unpivot(expr, asAlias, atAlias, byAlias, metas)
@@ -764,7 +763,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
 
     override fun visitLetBinding(node: Let.Binding, ctx: Ctx) = translate(node) { metas ->
         val expr = visitExpr(node.expr, ctx)
-        val name = stringToDefnid(node.asAlias.symbol)
+        val name = defnid(node.asAlias.symbol)
         letBinding(expr, name, metas)
     }
 
@@ -774,13 +773,13 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
             GroupBy.Strategy.PARTIAL -> groupPartial()
         }
         val keyList = groupKeyList(node.keys.translate<PartiqlAst.GroupKey>(ctx))
-        val groupAsAlias = node.asAlias?.let { stringToDefnid(it.symbol) }
+        val groupAsAlias = node.asAlias?.let { defnid(it.symbol) }
         groupBy(strategy, keyList, groupAsAlias, metas)
     }
 
     override fun visitGroupByKey(node: GroupBy.Key, ctx: Ctx) = translate(node) { metas ->
         val expr = visitExpr(node.expr, ctx)
-        val asAlias = node.asAlias?.let { stringToDefnid(it.symbol) }
+        val asAlias = node.asAlias?.let { defnid(it.symbol) }
         groupKey(expr, asAlias, metas)
     }
 
@@ -827,7 +826,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
             null -> null
         }
         val prefilter = node.prefilter?.let { visitExpr(it, ctx) }
-        val variable = node.variable?.let { stringToDefnid(it) }
+        val variable = node.variable?.let { defnid(it) }
         val quantifier = node.quantifier?.let { visitGraphMatchQuantifier(it, ctx) }
         val parts = node.parts.translate<PartiqlAst.GraphMatchPatternPart>(ctx)
         graphMatchPattern(restrictor, prefilter, variable, quantifier, parts, metas)
@@ -839,7 +838,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
     override fun visitGraphMatchPatternPartNode(node: GraphMatch.Pattern.Part.Node, ctx: Ctx) =
         translate(node) { metas ->
             val prefilter = node.prefilter?.let { visitExpr(it, ctx) }
-            val variable = node.variable?.let { stringToDefnid(it) }
+            val variable = node.variable?.let { defnid(it) }
             val label = node.label?.let { visitGraphMatchLabel(it, ctx) }
             node(prefilter, variable, label, metas)
         }
@@ -857,7 +856,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
             }
             val quantifier = node.quantifier?.let { visitGraphMatchQuantifier(it, ctx) }
             val prefilter = node.prefilter?.let { visitExpr(it, ctx) }
-            val variable = node.variable?.let { stringToDefnid(it) }
+            val variable = node.variable?.let { defnid(it) }
             val label = node.label?.let { visitGraphMatchLabel(it, ctx) }
             edge(direction, quantifier, prefilter, variable, label, metas)
         }
@@ -916,7 +915,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
 
     override fun visitGraphMatchLabelName(node: GraphMatch.Label.Name, ctx: Ctx) =
         translate(node) { metas ->
-            val id = stringToDefnid(node.name)
+            val id = defnid(node.name)
             graphLabelName(id, metas)
         }
 
@@ -954,7 +953,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
 
     override fun visitStatementDMLInsert(node: Statement.DML.Insert, ctx: Ctx) = translate(node) { metas ->
         val target = visitIdentifier(node.target, ctx)
-        val asAlias = node.asAlias?.let { stringToDefnid(it.symbol) }
+        val asAlias = node.asAlias?.let { defnid(it.symbol) }
         val values = visitExpr(node.values, ctx)
         val conflictAction = node.onConflict?.let { visitOnConflictAction(it.action, ctx) }
         val op = insert(target, asAlias, values, conflictAction)
@@ -978,7 +977,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
 
     override fun visitStatementDMLUpsert(node: Statement.DML.Upsert, ctx: Ctx) = translate(node) { metas ->
         val target = visitIdentifier(node.target, ctx)
-        val asAlias = node.asAlias?.let { stringToDefnid(it.symbol) }
+        val asAlias = node.asAlias?.let { defnid(it.symbol) }
         val values = visitExpr(node.values, ctx)
         val conflictAction = doUpdate(excluded())
         // UPSERT overloads legacy INSERT
@@ -988,7 +987,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
 
     override fun visitStatementDMLReplace(node: Statement.DML.Replace, ctx: Ctx) = translate(node) { metas ->
         val target = visitIdentifier(node.target, ctx)
-        val asAlias = node.asAlias?.let { stringToDefnid(it.symbol) }
+        val asAlias = node.asAlias?.let { defnid(it.symbol) }
         val values = visitExpr(node.values, ctx)
         val conflictAction = doReplace(excluded())
         // REPLACE overloads legacy INSERT
@@ -1034,9 +1033,9 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
 
     override fun visitStatementDMLDeleteTarget(node: Statement.DML.Delete.Target, ctx: Ctx) = translate(node) { metas ->
         val path = visitPathUnpack(node.path, ctx)
-        val asAlias = node.asAlias?.let { stringToDefnid(it.symbol) }
-        val atAlias = node.atAlias?.let { stringToDefnid(it.symbol) }
-        val byAlias = node.byAlias?.let { stringToDefnid(it.symbol) }
+        val asAlias = node.asAlias?.let { defnid(it.symbol) }
+        val atAlias = node.atAlias?.let { defnid(it.symbol) }
+        val byAlias = node.byAlias?.let { defnid(it.symbol) }
         scan(path, asAlias, atAlias, byAlias, metas)
     }
 
@@ -1084,7 +1083,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
         ctx: Ctx,
     ) = translate(node) { metas ->
         val target = visitIdentifier(node.target, ctx)
-        val asAlias = node.asAlias?.let { stringToDefnid(it.symbol) }
+        val asAlias = node.asAlias?.let { defnid(it.symbol) }
         val values = visitExpr(node.values, ctx)
         val conflictAction = node.onConflict?.let { visitOnConflictAction(it.action, ctx) }
         dmlOpList(insert(target, asAlias, values, conflictAction, metas))
@@ -1315,10 +1314,7 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
     override fun visitTypeAny(node: Type.Any, ctx: Ctx) = translate(node) { metas -> anyType(metas) }
 
     override fun visitTypeCustom(node: Type.Custom, ctx: Ctx) =
-        translate(node) { metas ->
-            val id = stringToDefnid(node.name)
-            customType(id.lowercase(), metas)
-        }
+        translate(node) { metas -> customType(defnid(node.name.lowercase()), metas) }
 
     /**
      * HELPERS
@@ -1329,14 +1325,6 @@ private class AstTranslator(val metas: Map<String, MetaContainer>) : AstBaseVisi
 
     private inline fun <reified T : PartiqlAst.PartiqlAstNode> visitOrNull(node: AstNode?, ctx: Ctx): T? =
         node?.let { visit(it, ctx) as T }
-
-    private fun stringToDefnid(str: String, metas: MetaContainer = emptyMetaContainer()): PartiqlAst.Defnid =
-        PartiqlAst.build {
-            defnid_(SymbolPrimitive(str, metas))
-        }
-
-    private fun PartiqlAst.Defnid.lowercase(): PartiqlAst.Defnid =
-        this.copy(symb = this.symb.copy(text = this.symb.text.lowercase()))
 
     private fun Identifier.CaseSensitivity.toLegacyCaseSensitivity() = when (this) {
         Identifier.CaseSensitivity.SENSITIVE -> PartiqlAst.CaseSensitivity.CaseSensitive()